+++ changelog.md
@@ -51,3 +51,4 @@
 - fixed bug with Bleeding tooltip not displaying proper value on server
 - fixed bug with Treasure Bag message appearing after changing worlds
 - fixed bug with Scroll animations not working properly on low frames per second
+- fixed bug with Enderium Shard Locator not working properly in the second hand (reported by @Ladocterus)